Splitted the Seat data and Seat functions to allow unit tests to get the Seats data
