+++ source/gamemap/GameMap.cpp
@@ -2032,6 +2032,9 @@ bool GameMap::addSeat(Seat *s)
         }
     }
     mSeats.push_back(s);
+    // We set the Seat color value
+    Ogre::ColourValue colorValue = ConfigManager::getSingleton().getColorFromId(s->getColorId());
+    s->setColorValue(colorValue);
 
     // Add the goals for all seats to this seat.
     for (auto& goal : mGoalsForAllSeats)
